[ML] Split CDataFrameAnalyzerTest class  (#761)
Split the CDataFrameAnalyzerTest class into outlier and training parts.
